add service to calculate distance

+++ math/src/main/java/pl/edu/pg/benchmarking/math/dto/DistanceRequest.java
@@ -10,43 +10,43 @@ public class DistanceRequest {
 
     private static final Logger LOG = Logger.getLogger(DistanceRequest.class.getName());
 
-    private Integer point1CoordinateX;
+    private Double point1CoordinateX;
 
-    private Integer point1CoordinateY;
+    private Double point1CoordinateY;
 
-    private Integer point2CoordinateX;
+    private Double point2CoordinateX;
 
-    private Integer point2CoordinateY;
+    private Double point2CoordinateY;
 
-    public Integer getPoint1CoordinateX() {
+    public Double getPoint1CoordinateX() {
         return point1CoordinateX;
     }
 
-    public Integer getPoint1CoordinateY() {
+    public Double getPoint1CoordinateY() {
         return point1CoordinateY;
     }
 
-    public Integer getPoint2CoordinateX() {
+    public Double getPoint2CoordinateX() {
         return point2CoordinateX;
     }
 
-    public Integer getPoint2CoordinateY() {
+    public Double getPoint2CoordinateY() {
         return point2CoordinateY;
     }
 
-    public void setPoint1CoordinateX(Integer point1CoordinateX) {
+    public void setPoint1CoordinateX(Double point1CoordinateX) {
         this.point1CoordinateX = point1CoordinateX;
     }
 
-    public void setPoint1CoordinateY(Integer point1CoordinateY) {
+    public void setPoint1CoordinateY(Double point1CoordinateY) {
         this.point1CoordinateY = point1CoordinateY;
     }
 
-    public void setPoint2CoordinateX(Integer point2CoordinateX) {
+    public void setPoint2CoordinateX(Double point2CoordinateX) {
         this.point2CoordinateX = point2CoordinateX;
     }
 
-    public void setPoint2CoordinateY(Integer point2CoordinateY) {
+    public void setPoint2CoordinateY(Double point2CoordinateY) {
         this.point2CoordinateY = point2CoordinateY;
     }
